Refactor codebase to use Modern C++20 features
- Replace `#define` constants in `source/app/definitions.h` with `constexpr int`.
- Remove fallback macro definitions in `source/map/position.h`.
- Replace `wxString::Printf` with `std::format` in `source/io/loaders/spr_loader.cpp`, `source/io/loaders/dat_loader.cpp`, and `source/rendering/utilities/tile_describer.cpp`.
- Modernize `NanoVGListBox` in `source/util/nanovg_listbox.cpp` to use `std::cmp_greater_equal` and `static_cast`.
- Replace C-style casts with `static_cast` in `source/util/nvg_utils.h`, `source/rendering/drawers/overlays/preview_drawer.cpp`, and `source/rendering/drawers/entities/creature_drawer.cpp`.
- Remove redundant casts in `source/ingame_preview/floor_visibility_calculator.cpp`.
- Replace `(ActionIdentifier)ACTION_SELECT` with `static_cast` in `source/live/live_action.cpp`.

+++ source/rendering/drawers/overlays/preview_drawer.cpp
@@ -68,7 +68,7 @@ void PreviewDrawer::draw(SpriteBatch& sprite_batch, MapCanvas* canvas, const Ren
 							b = b / 3 * 2;
 						}
 						if (tile->isHouseTile() && options.show_houses) {
-							if ((int)tile->getHouseID() == current_house_id) {
+							if (tile->getHouseID() == current_house_id) {
 								r /= 2;
 							} else {
 								r /= 2;
